Build scala-native from master branch as 0.4.0
This uses a locally build scala-native as a drop-in replacement for the official
scala-native 0.4.0 release.

The current 0.4.1-SNAPSHOT version from the master branch should be backward
compatible to 0.4.0, but currently is not because a method of the runtime
package has been removed some time ago.

Adding that method back makes libraries compiled for scala-native 0.4.0 able
to work with the SNAPSHOT build.