Also pass include dir for `cc` rule
The process_monitor implementation was changed from C++ to C in SN 0.5.7.

This led to the compilation in `release-full` mode to fail with an error that
the `sys/posix_sem.h` file could not be found targeting x86_64 darwin.

The file is available in `native/src/main/c/include` but that directory was only
passed with `-isystem` for C++ sources.