CI: Only restore specific zig cache
When restoring the cache, a restore-key is first checked whether there is
a (prefix) match for the specific key on the current branch, and only if there
is *no* match, the main branch will be checked.

This means a cache for zig 0.13 (say) will be restored for other zig versions
(say 0.14) and the cache will grow continuously. This is not what we want and
does not make any sense, since items in the zig cache are also dependent upon
the zig compiler.